Fix processing of files starting with normal comments
During the refactoring in v3.0.0 release, the way the comments are
extracted was reworked, causing files with headers to not be properly
parsed.

This fixes the parsing, clearing lines under processing. Otherwise
every line was being concatenated inside of the buffer.

The test project has been adjusted so it exercises this use case as
well.

Fixes: #25

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>